Experiment: do some traject processing in Rust
This is not ready for production.  It was an experiment to see how
much overhead it would be to pass a whole MarcXML record to rust
during a traject indexing routine.

It is slower, but not a lot slower -- and memoizing the Ruby method
more than made up for the overhead.

The results:
Original ruby implementation: 45288 records in 185.262 seconds; 244.5 records/second overall.
Passing the entire MarcXML document to rust, and reparsing it there: 45288 records in 194.183 seconds; 233.2 records/second overall.
Passing the entire MarcXML document to rust, and reparsing it there, but memoizing the results in Ruby: 45288 records in 180.621 seconds; 250.7 records/second overall.