[#2853] scsb:import:full rake task: choose the correct Dump if one is incorrect
Under the previous logic, if we had a Dump with a NULL generated_date,
it would always get chosen for a full SCSB re-index, even though it
is almost certainly not the Dump we would prefer.  For example,
yesterday, this logic caused an old Dump containing only SCSB private
records to be chosen over a newer Dump containing the correct records.

Closes #2853