AugmentTheSubject: use a Set to make subfield_a_match? faster
In a Vernier profile, 17% of all indexing time was spent in AugmentTheSubject#subfield_a_match?,
almost all of which was Array#include?

This commit switches from Array#include? to Set#include?, which is much faster. Vernier says that
this method now only take 0.6% of all indexing time, and Trajects statistics have increased from
240 records/second to 283 records/second locally.