Create a mock of the figgy mms ids translation map before the first indexing test
Previously we were creating it in config_spec.rb, which works well as long as
we run the tests alphabetically and don't have anything that comes before
config_spec.rb alphabetically that needs an IndexerService.

However, this PR adds a test file called cjk_spec.rb that needs an IndexerService,
so we move the fake translation map creation to be before running that test.

I initially tried using RSpec's when_first_matching_example_defined on the
:indexing tag, but it would not load WebMock within the config block, so I
just went with a before :all.