Remove unnecessary call from bibliographic holding availability endpoint
This affects the endpoint /bibliographic/:bib_id/holdings/:holding_id/availability,
which is used by the Requests form.

Prior to this commit, this endpoint made 2 or more requests to Alma:
* 1 to retrieve the MarcXML record along with requests and general
  availability data
* 1 or more to retrieve the item availability data

The first call was only used to retrieve the bib id -- something that
we already have because the user provided it in the request. Therefore,
we can eliminate this call and save an extra 500-1000ms each time we
load the request form (and be gentler on our API limits).

Co-authored-by: Christina Chortaria <christinach@users.noreply.github.com>
Co-authored-by: Mark Zelesky <mzelesky@users.noreply.github.com>