Make sure benchmark measures the correct thing
This benchmark was meant to benchmark the `genres` function,
but it spends much of its time parsing XML so that it can
actually run the function in question.

This commit moves the XML parsing out of the benchmark, so
that it can focus on measuring what we actually want to
measure.