Rewrite macrolanguage_codes in Rust
Also includes an optimization to the Rust ISO 639-3 code lookup.
Previously, it was slow for entries that are near the end of the
alphabet.  Switching to using a HashMap fixed this, and did not
slow down entries near the start of the alphabet.

Before (ruby implementation):
```
has macrolanguage, both are near the start of the alphabet
                          1.682M (± 3.2%) i/s  (594.57 ns/i) -      8.529M in   5.076489s
has no macrolanguage, middle of alphabet
                          1.950M (± 2.0%) i/s  (512.84 ns/i) -      9.747M in   5.000622s
has macrolanguage, near end of alphabet
                          1.690M (± 2.2%) i/s  (591.74 ns/i) -      8.448M in   5.001234s
```

After (rust implementation with new HashMap-based lookup):
```
has macrolanguage, both are near the start of the alphabet
                          2.833M (± 2.0%) i/s  (353.02 ns/i) -     14.394M in   5.083562s
has no macrolanguage, middle of alphabet
                          5.380M (± 1.6%) i/s  (185.87 ns/i) -     26.939M in   5.008385s
has macrolanguage, near end of alphabet
                          2.845M (± 1.6%) i/s  (351.46 ns/i) -     14.491M in   5.094399s
```